+++ frontend/components/collection/layouts/list/ListItem.tsx
@@ -57,7 +57,8 @@ export const ListItem = forwardRef<HTMLAnchorElement, ListItemProps>(
 
     return (
       <NextLink
-        href={`/collection/${collection.id}/article/${id}`}
+        href={`/?collectionId=${collection.id}&itemId=${id}`}
+        as={`/collection/${collection.id}/article/${id}`}
         ref={ref}
         className={cx(
           'block mr-3 border rounded hover:bg-purple-50 hover:dark:bg-gray-700',