+++ frontend/components/collection/list/CollectionList.tsx
@@ -11,7 +11,6 @@ import {
 } from '@shared';
 import { usePullToRefresh } from '@utils';
 import { CardItem, ListItem, MagazineItem, ExpandedItem } from '../layouts';
-import { RefreshIndicator } from './RefreshIndicator';
 import { GroupHeader } from './GroupHeader';
 import { PanesLayout } from '../types';
 
@@ -145,7 +144,7 @@ export const CollectionList: React.FC<CollectionListProps & BoxProps> = (
       data-test-layout={layout}
       {...rest}
     >
-      <RefreshIndicator isRefreshing={isRefreshing} />
+      {/* <RefreshIndicator isRefreshing={isRefreshing} /> */}
       {items.type === 'list' ? (
         <Virtuoso
           key="flat"