+++ frontend/pages/collection/[collectionId]/article/[itemId].page.tsx
@@ -1,23 +1,23 @@
 import React from 'react';
-import type { NextPage } from 'next';
 import Head from 'next/head';
-import { useRouter } from 'next/router';
-import { getNumber } from '@utils/router';
-import { useArticleDetails } from '@features/Article/queries';
+import { NextPageWithLayout } from '@pages/types';
+import { Panes } from '@features/Panes';
+import { usePageTitle } from '@pages/usePageTitle';
 
-const ItemPage: NextPage = () => {
-  const router = useRouter();
-  const collectionId = getNumber(router.query?.collectionId);
-  const itemId = getNumber(router.query?.itemId);
-  const { data: { title } = {} } = useArticleDetails(collectionId!, itemId!);
+const StandaloneItemPage: NextPageWithLayout = () => {
+  const title = usePageTitle();
 
   return (
     <>
       <Head>
-        <title>{title ? `${title} | Orpington News` : 'Orpington News'}</title>
+        <title>{title}</title>
       </Head>
     </>
   );
 };
 
-export default ItemPage;
+StandaloneItemPage.getLayout = (page) => {
+  return <Panes standaloneArticle>{page}</Panes>;
+};
+
+export default StandaloneItemPage;